Extract SearchInput to component
