+++ app/screens/select-station/select-station-screen.tsx
@@ -1,12 +1,12 @@
 import React, { useMemo, useState } from "react"
 import { observer } from "mobx-react-lite"
-import { FlatList, LayoutAnimation, View, TextInput, TextStyle, ViewStyle, Pressable, I18nManager } from "react-native"
+import { FlatList, View, TextStyle, ViewStyle, Pressable } from "react-native"
 import { Screen, Text, StationCard } from "../../components"
 import { useStores } from "../../models"
 import { SelectStationScreenProps } from "../../navigators/main-navigator"
-import { color, spacing, typography } from "../../theme"
+import { color, spacing } from "../../theme"
 import { useStations } from "../../data/stations"
-import { translate } from "../../i18n"
+import { SearchInput } from "./search-input"
 
 // #region styles
 const ROOT: ViewStyle = {
@@ -20,17 +20,6 @@ const SEARCH_BAR_WRAPPER: ViewStyle = {
   paddingHorizontal: spacing[3],
   marginBottom: spacing[3],
   marginTop: spacing[2],
-}
-
-const SEARCH_BAR: TextStyle = {
-  flex: 1,
-  padding: spacing[3],
-
-  textAlign: I18nManager.isRTL ? "right" : "left",
-  fontFamily: typography.primary,
-  borderRadius: 8,
-  color: color.text,
-  backgroundColor: color.background,
 }
 
 const CANCEL_LINK: TextStyle = {
@@ -74,29 +63,7 @@ export const SelectStationScreen = observer(function SelectStationScreen({ navig
   return (
     <Screen style={ROOT} preset="fixed" unsafe={false}>
       <View style={SEARCH_BAR_WRAPPER}>
-        <TextInput
-          style={SEARCH_BAR}
-          placeholder={translate("selectStation.placeholder")}
-          placeholderTextColor={color.dim}
-          onChangeText={(text) => {
-            LayoutAnimation.configureNext({
-              duration: 400,
-              create: {
-                type: LayoutAnimation.Types.spring,
-                property: LayoutAnimation.Properties.opacity,
-                springDamping: 1,
-              },
-              delete: {
-                type: LayoutAnimation.Types.spring,
-                property: LayoutAnimation.Properties.opacity,
-                springDamping: 1,
-              },
-            })
-            setSearchTerm(text)
-          }}
-          autoFocus={true}
-          autoCorrect={false}
-        />
+        <SearchInput setSearchTerm={setSearchTerm} />
         <Pressable onPress={() => navigation.navigate("planner")}>
           <Text style={CANCEL_LINK} tx="common.cancel" />
         </Pressable>
